+++ R/classes.R
@@ -467,13 +467,14 @@ updateGiottoObject <- function(gobject) {
     if (!is.null(attr(gobject, "OS_platform"))) {
         attr(gobject, "OS_platform") <- NULL
     }
-    if (is.null(attr(gobject, "versions"))) { # apply default version 0
+    if (is.null(attr(gobject, "versions"))) { # apply default version 0.0.0
         attr(gobject, "versions") <- .versions_info()
-        gobject@versions$gclass <- 0 # untracked
+        gobject@versions$gclass <- "0.0.0" # untracked
     }
 
     # warn if gobject newer than package
-    if (as.character(.gversion(gobject)) > as.character(packageVersion("GiottoClass"))) {
+    if (.gversion(gobject) >
+        numeric_version(packageVersion("GiottoClass"))) {
         warning(
             call. = FALSE,
             sprintf(
@@ -486,18 +487,23 @@ updateGiottoObject <- function(gobject) {
 
     # [version-based updates] -------------------------------------------------#
 
-    if (.gversion(gobject) < "0.5.1") {
+    if (.gversion(gobject) < numeric_version("0.5.1")) {
         gobject <- .update_source_slot(gobject)
     }
 
     # GiottoClass 0.3.0 removes @largeImages slot
-    if (.gversion(gobject) < "0.3.0") {
+    if (.gversion(gobject) < numeric_version("0.3.0")) {
         gobject <- .update_image_slot(gobject)
     }
 
     # GiottoClass 0.1.2 image updates moved here
     # TODO remove in future update
     gobject@images <- lapply(gobject@images, .update_giotto_image)
+
+    # GiottoClass 0.4.12 adds @misc slot
+    if (.gversion(gobject) < numeric_version("0.4.12")) {
+        attr(gobject, "misc") <- list()
+    }
 
     # -------------------------------------------------------------------------#
 
@@ -636,6 +642,7 @@ updateGiottoObject <- function(gobject) {
 #' @slot h5_file path to h5 file
 #' @slot source path to source directory backing the `giotto` object. If `""`,
 #' the object is in memory.
+#' @slot misc miscellaneous or unstructured data
 #' @details
 #'
 #' \[**initialize**\]
@@ -688,7 +695,8 @@ giotto <- setClass(
         join_info = "ANY",
         multiomics = "ANY",
         h5_file = "ANY",
-        source = "ANY"
+        source = "ANY",
+        misc = "list"
         # mirai = 'list'
     ),
     prototype = list(
@@ -714,7 +722,8 @@ giotto <- setClass(
         join_info = NULL,
         multiomics = NULL,
         h5_file = NULL,
-        source = ""
+        source = "",
+        misc = list()
         # mirai = list()
     )
 
@@ -765,7 +774,8 @@ setClass(
         versions = "ANY",
         join_info = "ANY",
         multiomics = "ANY",
-        h5_file = "ANY"
+        h5_file = "ANY",
+        misc = "list"
     ),
     prototype = list(
         packed_spatial_info = NULL,
@@ -789,7 +799,8 @@ setClass(
         versions = NULL,
         join_info = NULL,
         multiomics = NULL,
-        h5_file = NULL
+        h5_file = NULL,
+        misc = list()
     )
 )