+++ R/create.R
@@ -3160,7 +3160,7 @@ createGiottoImage <- function(gobject = NULL,
             if (transf == "flip_x_axis") {
                 mg_object <- magick::image_flop(mg_object)
             } else if (transf == "flip_y_axis") {
-                mg_object <- magick::image_flop(mg_object)
+                mg_object <- magick::image_flip(mg_object)
             } else {
                 wrap_msg(transf, " is not a supported transformation, see
                     details")