+++ NEWS.md
@@ -11,6 +11,14 @@
 - `overlaps()` will now properly find image overlaps
 
 
+# GiottoClass 0.4.12 (TBD)
+
+## new
+- `misc` slot for storing unstructured data
+
+## enhancements
+- escape hatch for gobject initialize checking. Set option `"giotto.init_check_severity"` to `"stop"` (default), or `"warning"` depending on needs.
+
 # GiottoClass 0.4.10 (2025/09/30)
 
 ## bug fixes